requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-shortcuts==1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-dialogflow==2.31.0
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
gunicorn==22.0.0
idna==3.8
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.9.0
pytz==2024.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
urllib3==2.2.2
